+++ packages/quantic/force-app/main/default/lwc/quanticFacet/e2e/pageObject.ts
@@ -1,8 +1,4 @@
-import type {Locator, Page, Request, Response} from '@playwright/test';
-import {
-  isUaCustomEvent,
-  isUaSearchEvent,
-} from '../../../../../../playwright/utils/requests';
+import type {Locator, Page, Response} from '@playwright/test';
 
 export class FacetObject {
   constructor(public page: Page) {
@@ -63,124 +59,6 @@ export class FacetObject {
 
   async clickOnClearSelectionButton(): Promise<void> {
     await this.clearSelectionButton.click();
-  }
-
-  async waitForFacetUaSearchEvent(
-    actionCause: string,
-    customChecker?: Function
-  ): Promise<Request> {
-    const uaRequest = this.page.waitForRequest((request) => {
-      if (isUaSearchEvent(request)) {
-        const requestBody = request.postDataJSON?.();
-        const {customData} = requestBody;
-
-        const expectedFields: Record<string, any> = {
-          actionCause: actionCause,
-        };
-        const matchesExpectedFields = Object.keys(expectedFields).every(
-          (key) => requestBody?.[key] === expectedFields[key]
-        );
-
-        return (
-          matchesExpectedFields &&
-          (customChecker ? customChecker(customData) : true)
-        );
-      }
-      return false;
-    });
-    return uaRequest;
-  }
-
-  async waitForFacetUaCustomEvent(
-    actionCause: string,
-    customChecker?: Function
-  ): Promise<Request> {
-    const uaRequest = this.page.waitForRequest((request) => {
-      if (isUaCustomEvent(request)) {
-        const requestBody = request.postDataJSON?.();
-        const {customData} = requestBody;
-
-        const expectedFields: Record<string, any> = {
-          eventValue: actionCause,
-          eventType: 'facet',
-        };
-        const matchesExpectedFields = Object.keys(expectedFields).every(
-          (key) => requestBody?.[key] === expectedFields[key]
-        );
-
-        return (
-          matchesExpectedFields &&
-          (customChecker ? customChecker(customData) : true)
-        );
-      }
-      return false;
-    });
-    return uaRequest;
-  }
-
-  async waitForFacetSelectUaAnalytics(
-    expectedFields: Record<string, any>
-  ): Promise<Request> {
-    return this.waitForFacetUaSearchEvent(
-      'facetSelect',
-      (customData: Record<string, any>) => {
-        return Object.keys(expectedFields).every(
-          (key) => customData?.[key] === expectedFields[key]
-        );
-      }
-    );
-  }
-
-  async waitForFacetDeselectUaAnalytics(
-    expectedFields: Record<string, any>
-  ): Promise<Request> {
-    return this.waitForFacetUaSearchEvent(
-      'facetDeselect',
-      (customData: Record<string, any>) => {
-        return Object.keys(expectedFields).every(
-          (key) => customData?.[key] === expectedFields[key]
-        );
-      }
-    );
-  }
-
-  async waitForFacetClearAllUaAnalytics(
-    expectedFields: Record<string, any>
-  ): Promise<Request> {
-    return this.waitForFacetUaSearchEvent(
-      'facetClearAll',
-      (customData: Record<string, any>) => {
-        return Object.keys(expectedFields).every(
-          (key) => customData?.[key] === expectedFields[key]
-        );
-      }
-    );
-  }
-
-  async waitForShowMoreFacetResultsUaAnalytics(
-    expectedFields: Record<string, any>
-  ): Promise<Request> {
-    return this.waitForFacetUaCustomEvent(
-      'showMoreFacetResults',
-      (customData: Record<string, any>) => {
-        return Object.keys(expectedFields).every(
-          (key) => customData?.[key] === expectedFields[key]
-        );
-      }
-    );
-  }
-
-  async waitForShowLessFacetResultsUaAnalytics(
-    expectedFields: Record<string, any>
-  ): Promise<Request> {
-    return this.waitForFacetUaCustomEvent(
-      'showLessFacetResults',
-      (customData: Record<string, any>) => {
-        return Object.keys(expectedFields).every(
-          (key) => customData?.[key] === expectedFields[key]
-        );
-      }
-    );
   }
 
   extractDataFromSearchResponse(response: Response) {
